webapi/webapi-presentation-w3c-tests/presentation/PresentationRequest_reconnect.html

<!DOCTYPE html>
<!--
Copyright (c) 2015 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-->

<meta charset="utf-8">
<title>PresentationRequest.reconnect()</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Chunyan Wang" href="mailto:chunyanx.wang@intel.com">
<link rel="help" href="http://w3c.github.io/presentation-api/#dom-presentationrequest-reconnect">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
promise_test(function(t) {
  var presId;
  var url = "display/presentation.html";
  var request = new PresentationRequest(url);
  return request.start()
    .then(function(connection) {
      assert_true(connection instanceof PresentationConnection);
      presId = connection.id;
      var closed = new Promise(function(resolve) {
        connection.onclosed = t.step_func(function() {
          assert_equals(this.state, "closed");
          resolve();
        });
      });
      var connected = new Promise(function(resolve) {
        connection.onconnected = t.step_func(function() {
          this.close();
          resolve();
        });
      });
      return Promise.all([closed, connected]);
    })
    .then(function() {
      return request.reconnect(presId);
    })
    .then(function(reconnection) {
      assert_true(reconnection instanceof PresentationConnection);
      assert_equals(reconnection.id, presId);
    })
    .catch(function(ex) {
      assert_unreached(ex.name + ":" + ex.message);
    });
}, "Check that PresentationRequest.reconnect() can reconnect the presentation displays with known presentationId.");
</script>
